data_preprocessing/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tab2graph"
version = "0.1.0"
description = "A unified framework for tabular and graph machine learning with optional DGL support"
readme = "README.md"
license = {text = "Apache-2.0"}
authors = [
    {name = "AWS Shanghai AI Lab"}
]
maintainers = [
    {name = "AWS Shanghai AI Lab"}
]
keywords = ["machine learning", "tabular data", "graph neural networks", "benchmark"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"

# Core dependencies (always required)
dependencies = [
    # Core data processing
    "numpy",
    "pandas", 
    "pydantic==1.10.12",
    "pyyaml",
    "requests[security]",
    "scikit-learn",
    "scipy",
    "tqdm",
    "typer",
    "psutil",
    
    # PyTorch ecosystem (required for tabular ML)
    "torch",
    "torchmetrics",
    
    # Tabular ML dependencies
    "autogluon-tabular[all]",
    
    # Database dependencies
    "sqlalchemy",
    "duckdb", 
    "sqlglot==27.11.0",
    "sql_formatter",
    
    # Logging and monitoring
    "wandb",
    
    # Plotting
    "matplotlib",
    
    # Storage
    "s3fs",
    "boto3",
    
    # Feature engineering
    "featuretools",
    
    # NLP (for text features)
    "tokenizers==0.22.0",
    "transformers",
    "tiktoken",
    "gensim",

    # h5 dataset
    "h5py",
]

[project.optional-dependencies]
# Graph ML dependencies (optional)
graph = [
    "dgl>=2.1a240205",
    "networkx",
    "ogb",
    "einops",
]

# Development dependencies
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "isort",
    "flake8",
    "mypy",
]

# All optional dependencies
all = [
    "tab2graph[graph,dev]"
]

[project.scripts]
# CLI entry points
tab2graph = "tab2graph.main:app"

[project.urls]
Homepage = "https://github.com/awslabs/multi-table-benchmark"
Repository = "https://github.com/awslabs/multi-table-benchmark"
Issues = "https://github.com/awslabs/multi-table-benchmark/issues"
Documentation = "https://github.com/awslabs/multi-table-benchmark/blob/main/README.md"

[tool.setuptools.packages.find]
include = ["tab2graph*", "dbinfer_bench*"]
exclude = ["tests*", "docs*", "examples*", "_backup_examples*", "wandb*"]

[tool.setuptools.package-data]
"dbinfer_bench" = ["download_config.yaml"]
"tab2graph" = ["**/*.yaml", "**/*.yml"]

# Black configuration
[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | build
  | dist
  | wandb
  | _backup_examples
)/
'''

# isort configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["wandb/*", "_backup_examples/*"]

# pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "-ra",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "requires_dgl: marks tests that require DGL to be installed",
]

# MyPy configuration
[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
exclude = [
    "wandb/",
    "_backup_examples/",
    "build/",
    "dist/",
]
